add upload item component
