+++ app/layout/projects/all/toolbar/upload-btn/legacy-upload-modal/upload-files/upload-item/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
